refactored ElasticsearchQuery removing all mapping logic into the new ESMapper-class
